refactor: split vision candidate helpers
